refactor connection logic
